tests/pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=7.0.0",
    "psycopg2-binary>=2.9.0",
    "pytest-asyncio>=0.21.0",
]
